+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Upgraded SWIFT registry provider to July '23 release 95, which adds Falkland Islands (FK).
+- Upgraded Wikipedia registry provider, which adds North Macedonia (MK).
 - [#144](https://github.com/skwasjer/IbanNet/pull/144) Add Catalan localization.
 - [#138](https://github.com/skwasjer/IbanNet/pull/138) Implement `IParsable<TSelf>` on Iban type.
 - In `IbanParser` internals, use `ReadOnlySpan<T>`.